Abstractify `HUDHitErrorMeter` to isolate color determination
